+++ app/Models/Basket.php
@@ -6,6 +6,8 @@ use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+use Illuminate\Database\Eloquent\Relations\HasManyThrough;
 
 class Basket extends MainModel
 {
@@ -25,5 +27,10 @@ class Basket extends MainModel
     public function basketItems(): HasMany
     {
         return $this->hasMany(BasketItem::class, 'basket_id', 'id');
+    }
+
+    public function products(): BelongsToMany
+    {
+        return $this->belongsToMany(Product::class,'basket_items');
     }
 }