+++ resources/views/partials/_single_product.blade.php
@@ -4,11 +4,8 @@
             <img src="https://via.placeholder.com/335x335" alt="#">
 
             @include('products.partials._product_badge',['product' => $product])
-
-
             <div class="button">
-                <a href="{{route('basket.add-to-basket',$product)}}" class="btn"><i
-                        class="lni lni-cart"></i> Add to Cart</a>
+                @include('products.partials._add_to_cart')
             </div>
         </div>
         <div class="product-info">
@@ -16,7 +13,7 @@
             <h4 class="title">
                 <a href="{{route('products.show',$product)}}">{{$product->name}}</a>
             </h4>
-           @include('partials._review',['product' => $product])
+            @include('partials._review',['product' => $product])
             @include('partials._price',['product' => $product])
         </div>
     </div>