+++ resources/views/products/partials/_add_to_cart.blade.php
@@ -1,0 +1,8 @@
+@if(!$basket->products->contains($product))
+    <a href="{{route('basket.add-to-basket',$product)}}" class="btn" style="{{$style ?? ''}}">
+        <i class="lni lni-cart"></i>
+        Add to Cart
+    </a>
+@else
+    <a href="{{route('basket.index')}}" class="btn" style="{{$style ?? ''}}">Go to basket</a>
+@endif