+++ resources/views/products/show.blade.php
@@ -79,7 +79,7 @@
                                 <div class="row align-items-end">
                                     <div class="col-lg-4 col-md-4 col-12">
                                         <div class="button cart-button">
-                                            <a href="{{route('basket.add-to-basket',$product)}}" class="btn"  style="width: 100%;padding:13px 30px">Add to Cart</a>
+                                           @include('products.partials._add_to_cart',['style' => 'width: 100%;padding:13px 30px'])
                                         </div>
                                     </div>
                                     <div class="col-lg-4 col-md-4 col-12">